refactor: containers 폴더 components 하위로 이동
