+++ src/components/Container/PostContainer/PostContainer.module.scss
@@ -1,6 +1,6 @@
-@import '../../styles/colors';
-@import '../../styles/mixins';
-@import '../../styles/sizes';
+@import '@styles/colors';
+@import '@styles/mixins';
+@import '@styles/sizes';
 
 .container {
   max-width: 1024px;
